+++ lib/bdaya_flutter_common.dart
@@ -8,7 +8,6 @@ export 'package:logging/logging.dart';
 export 'package:get_it/get_it.dart';
 export 'package:tuple/tuple.dart';
 export 'package:rxdart/rxdart.dart';
-export 'package:get_it/get_it.dart';
 export 'package:injectable/injectable.dart';
 //Self exports
 export 'src/helpers/_exports.dart';