+++ lib/src/extensions/streams.dart
@@ -1,7 +1,8 @@
 import 'package:bdaya_flutter_common/bdaya_flutter_common.dart';
 
 extension LogStreamsExt<T> on Stream<T> {
-  Stream<T> wrapWithArea(BdayaRxLoadableArea area, [Logger? logger, String? message]) {
+  Stream<T> wrapWithArea(BdayaRxLoadableArea area,
+      [Logger? logger, String? message]) {
     return doOnListen(() {
       area.startLoading();
       logger?.finest('[${area.$.displayName}] Started loading');
@@ -11,13 +12,15 @@ extension LogStreamsExt<T> on Stream<T> {
         logger?.finest('[${area.$.displayName}] Stopped loading successfully');
       },
     ).doOnError(
-      (error, stackTrace) => area.stopLoadingError(error, stackTrace, logger, message),
+      (error, stackTrace) =>
+          area.stopLoadingError(error, stackTrace, logger, message),
     );
   }
 }
 
 extension LogFutureExt<T> on Future<T> {
-  Future<T> wrapWithArea(BdayaRxLoadableArea area, [Logger? logger, String? message]) async {
+  Future<T> wrapWithArea(BdayaRxLoadableArea area,
+      [Logger? logger, String? message]) async {
     try {
       area.startLoading();
       logger?.finest('[${area.$.displayName}] Started loading');