+++ lib/src/helpers/combined_controllers.dart
@@ -5,7 +5,13 @@ import 'controller/is_loading_mixin.dart';
 import 'controller/loggable_mixin.dart';
 
 /// For all controllers (Dialogs + Pages)
-abstract class BdayaCombinedController with BdayaLoggableMixin, BdayaLifeCycleMixin, BdayaIsLoadingMixin, BdayaDisposeHandlerMixin {}
+abstract class BdayaCombinedController
+    with
+        BdayaLoggableMixin,
+        BdayaLifeCycleMixin,
+        BdayaIsLoadingMixin,
+        BdayaDisposeHandlerMixin {}
 
 /// For page controllers that require route information
-abstract class BdayaCombinedRouteController extends BdayaCombinedController with BdayaRouterListenableMixin {}
+abstract class BdayaCombinedRouteController extends BdayaCombinedController
+    with BdayaRouterListenableMixin {}